Report the CPU emulator’s error output if it fails
Now that we’re capturing standard error from the emulator into a string,
we can use it as `assert`’s failure message. This will make a broken
acceptance test fail with a message like “Comparison failure at line 2”,
which is a bit more useful than the default “Expected false to be
truthy”.

We could provide an even more useful failure message by generating a
diff between the `.out` and `.cmp` files ourselves, but the error
message we already have is better than nothing.