Include a diff in the acceptance test failure message
The `.out` file generated during an acceptance test gets automatically
deleted at the end, which can make it difficult to debug an acceptance
test failure. This commit generates a diff [0] between the `.out` and
`.cmp` files and includes it in the assertion message to make it easier
to understand why the test is failing.

For example:

```
  1) Failure:
VMTranslatorAcceptanceTest#test_acceptance_SimpleAdd [test/acceptance/vm_translator_test.rb:47]:
Comparison failure at line 2
--- expected
+++ actual
@@ -1,3 +1,3 @@
 "|  RAM[0]  | RAM[256] |
-|     257  |      16  |
+|     257  |      15  |
 "
```

[0] http://docs.seattlerb.org/minitest/Minitest/Assertions.html#method-i-diff