Automatically generate an acceptance test for each script
This copies the technique used for the integration tests in #12: we
iterate over each of the `test/fixtures/*.tst` files and use
`define_method` to generate a test for each of them individually.

Now we can add acceptance tests for other `examples/*.vm` files simply
by copying the appropriate `.tst` and `.cmp` files into `test/fixtures`,
which we‘ll do in the next commit.